@@ -1,13 +1,11 @@
 package com.demotxt.droidsrce.homedashboard;
 
-import android.animation.Animator;
 import android.content.Intent;
-import android.os.Handler;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.CardView;
+import androidx.cardview.widget.CardView;
 import android.view.View;
-import android.view.ViewAnimationUtils;
 import android.widget.LinearLayout;
 
 
